# refact(raft): concurrent DB updates and GQL rebuild on db reload