--- a/cluster/schema/manager.go
+++ b/cluster/schema/manager.go
@@ -123,7 +123,7 @@ func (s *SchemaManager) ReloadDBFromSchema() {
 		cs[i] = command.UpdateClassRequest{Class: &v.Class, State: shardingState}
 		i++
 	}
-
+	s.db.TriggerSchemaUpdateCallbacks()
 	s.log.Info("reload local db: update schema ...")
 	s.db.ReloadLocalDB(context.Background(), cs)
 }