--- a/cluster/schema/manager.go
+++ b/cluster/schema/manager.go
@@ -21,9 +21,10 @@ import (
 	"github.com/hashicorp/raft"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/sirupsen/logrus"
+	gproto "google.golang.org/protobuf/proto"
+
 	"github.com/weaviate/weaviate/cluster/proto/api"
 	command "github.com/weaviate/weaviate/cluster/proto/api"
-	gproto "google.golang.org/protobuf/proto"
 )
 
 var (
@@ -363,14 +364,15 @@ func (s *SchemaManager) apply(op applyOp) error {
 		return fmt.Errorf("%w: %s: %w", ErrSchema, op.op, err)
 	}
 
-	if op.enableSchemaCallback {
-		s.db.TriggerSchemaUpdateCallbacks()
-	}
-
 	if !op.schemaOnly {
 		if err := op.updateStore(); err != nil {
 			return fmt.Errorf("%w: %s: %w", errDB, op.op, err)
 		}
+	}
+
+	// Always trigger the schema callback last
+	if op.enableSchemaCallback {
+		s.db.TriggerSchemaUpdateCallbacks()
 	}
 
 	return nil