--- a/usecases/schema/executor.go
+++ b/usecases/schema/executor.go
@@ -15,9 +15,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"runtime"
 	"sync"
 
 	"github.com/sirupsen/logrus"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/weaviate/weaviate/cluster/proto/api"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
@@ -56,24 +58,36 @@ func (e *executor) Open(ctx context.Context) error {
 func (e *executor) ReloadLocalDB(ctx context.Context, all []api.UpdateClassRequest) error {
 	cs := make([]*models.Class, len(all))
 
+	g, ctx := errgroup.WithContext(ctx)
+	g.SetLimit(runtime.GOMAXPROCS(0) * 2)
+
 	var errList error
-	wg := sync.WaitGroup{}
-	wg.Add(len(all))
+	var errMutex sync.Mutex
+
 	for i, u := range all {
-		i := i
-		u := u
-		enterrors.GoWrapper(func() {
-			defer wg.Done()
+		i, u := i, u
+
+		g.Go(func() error {
 			e.logger.WithField("index", u.Class.Class).Info("reload local index")
 			cs[i] = u.Class
+
 			if err := e.migrator.UpdateIndex(ctx, u.Class, u.State); err != nil {
 				e.logger.WithField("index", u.Class.Class).WithError(err).Error("failed to reload local index")
-				errList = errors.Join(fmt.Errorf("failed to reload local index %q: %w", i, err))
+				err := fmt.Errorf("failed to reload local index %q: %w", i, err)
+
+				errMutex.Lock()
+				errList = errors.Join(errList, err)
+				errMutex.Unlock()
 			}
-		}, e.logger)
+			return nil
+		})
 	}
+
+	if err := g.Wait(); err != nil {
+		return err
+	}
+
 	e.TriggerSchemaUpdateCallbacks()
-	wg.Wait()
 	return errList
 }