### What's being changed:
**current implementation**  
- loads the index sequentially 
- schema call backs sequentially (like rebuilding GQL)
**new with this PR change** 
- load concurrently 
- rebuild concurrently 

**Results** 
1.25 (**Current**)
- Inserting 500 classes took `116.23s`
- From live to ready after restart took `151000ms`

this PR branch  (**new**)
- Inserting 500 classes took `85.59s`
- From live to ready after restart took `9000ms`

chaos pipeline https://github.com/weaviate/weaviate-chaos-engineering/actions/runs/13138254690
e2e tests pipeline https://github.com/weaviate/weaviate-e2e-tests/actions/runs/13138256389


another lates run 
chaos https://github.com/weaviate/weaviate-chaos-engineering/actions/runs/13176904222
e2e https://github.com/weaviate/weaviate-e2e-tests/actions/runs/13176901854 
### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.

<!-- Uncomment the following section if this PR requires changes in related projects (e.g., documentation, client libraries).

GitHub actions will automatically create an issue in the corresponding repository for each checked box below. (See `.github/workflows/create-cross-functional-issues.yml`)

### Cross-functional impact

- [ ] This change requires public documentation (weaviate-io) to be updated. Check the box to automatically create a corresponding issue.
- Does it require a change in the client libraries? If yes, please check the boxes for the affected client libraries.
    - [ ] Python (weaviate-python-client)
    - [ ] JavaScript/TypeScript (typescript-client)
    - [ ] Go (weaviate-go-client)
    - [ ] Java (java-client)

-->